ROTTEN TOMATOES: CLASSIFYING MOVIE REVIEW TYPE BASED ON REVIEW SENTIMENT
This code makes use of text analytics to conduct a sentiment analysis of understanding the sentiment behind reviews and classify it as Fresh or Rotten based on the sentiment. This code encompasses multiple data and pre processing steps geared to develop a clean dataset suitable to conduct a text analytics project. The model behind this code is logistic regression